requirements.txt
# ATLAS Requirements
# Python dependencies for ATLAS AI Assistant

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Telegram Bot
python-telegram-bot==21.0

# Database
supabase>=2.8.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# OpenAI & Embeddings
openai==1.10.0
tiktoken==0.5.2

# Claude (Anthropic)
anthropic>=0.18.0

# Vector & ML
numpy==1.26.3

# Environment & Configuration
python-dotenv==1.0.1

# Utilities
pydantic-settings==2.1.0

# Logging (optional enhancements)
python-json-logger==2.0.7

# Development & Testing (optional)
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.1.14

# Monitoring (optional)
prometheus-client==0.19.0

# Rate Limiting (optional)
slowapi==0.1.9

# Additional utilities
requests==2.31.0
aiofiles==23.2.1

# Background task scheduling
APScheduler==3.10.4
pytz==2024.1
